pyproject.toml
[project]
name = "fastapibeyondcrud"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.11",
    "passlib>=1.7.4",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "redis>=4.2.0rc1",
    "sqlmodel>=0.0.24",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
]
